Don't create [Content_Types].old in appx files
In addContentType in APPXFile.java, each time a new content type is
seen it renames [Content_Types].xml to [Content_Types].old and then
writes a new [Content_Types].xml.

I don't know why this would be necessary and it causes the following
error when using makemsix (the Microsoft tool from
https://github.com/microsoft/msix-packaging):

    Payload file not described in AppxBlockMap.xml

The payload file it's complaining about is [Content_Types].old.

This patch removes the call to renameEntry so [Content_Types].old is
never created.